+++ src/pages/cahoots/index.tsx
@@ -7,7 +7,7 @@ import DeadlineCarousel from '@/components/cahoot/DeadlineCarousel';
 import List from '@/components/cahoot/List';
 import Recap from '@/components/cahoot/Recap';
 import Layout from '@/components/common/Layout';
-import Interests from '@/components/market/Interests';
+import Interests from '@/components/common/Interests';
 import wrapper from '@/store';
 import { ErrorBoundary } from '@sentry/nextjs';
 import { dehydrate, QueryClient } from '@tanstack/react-query';
